+++ src/frontend/web_application/src/scenes/Discussion/components/InstantMessage/index.jsx
@@ -2,16 +2,22 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import Moment from 'react-moment';
 import classNames from 'classnames';
-import { withI18n } from '@lingui/react';
+import { withI18n, Trans } from '@lingui/react';
 import { withScrollTarget } from '../../../../modules/scroll';
-import { isMessageFromUser, getAuthor, isUserRecipient, getRecipientsExceptUser, getRecipients } from '../../../../services/message';
+import { withPush } from '../../../../modules/routing';
 import { getAveragePIMessage, getPiClass } from '../../../../modules/pi';
 import { AuthorAvatarLetter } from '../../../../modules/avatar';
-import { Icon, TextBlock } from '../../../../components';
 import { LockedMessage } from '../../../../modules/encryption';
+import { Button, Confirm, Icon, TextBlock } from '../../../../components';
+import { isMessageFromUser, getAuthor, isUserRecipient, getRecipientsExceptUser, getRecipients } from '../../../../services/message';
 import MessagePi from '../MessagePi';
+import TagList from '../TagList';
+import { replyHandler } from '../../services/replyHandler';
+import { messageDeleteHandler } from '../../services/messageDeleteHandler';
+import { toggleMarkAsReadHandler } from '../../services/toggleMarkAsReadHandler';
 
 import './style.scss';
+import './instant-message-aside.scss';
 import './instant-message-author.scss';
 import './instant-message-participants.scss';
 
@@ -25,15 +31,19 @@ const PROTOCOL_ICONS = {
 
 @withI18n()
 @withScrollTarget()
+@withPush()
 class InstantMessage extends PureComponent {
   static propTypes = {
     message: PropTypes.shape({}).isRequired,
     i18n: PropTypes.shape({}).isRequired,
     // XXX: No UI for that
-    // onMessageRead: PropTypes.func,
-    // onMessageUnread: PropTypes.func,
-    // onDeleteMessage: PropTypes.func,
-    // noInteractions: PropTypes.bool,
+    noInteractions: PropTypes.bool,
+    onMessageRead: PropTypes.func,
+    onMessageUnread: PropTypes.func,
+    onMessageDelete: PropTypes.func,
+    onOpenTags: PropTypes.func.isRequired,
+    onReply: PropTypes.func.isRequired,
+    push: PropTypes.func.isRequired,
     user: PropTypes.shape({}).isRequired,
     scrollTarget: PropTypes.shape({ forwardRef: PropTypes.func }).isRequired,
     isLocked: PropTypes.bool.isRequired,
@@ -42,6 +52,10 @@ class InstantMessage extends PureComponent {
 
   static defaultProps = {
     encryptionStatus: undefined,
+    onMessageRead: () => {},
+    onMessageUnread: () => {},
+    onMessageDelete: () => {},
+    noInteractions: false,
   };
 
   getProtocolIconType = ({ protocol }) => PROTOCOL_ICONS[protocol] || 'comment';
@@ -76,6 +90,56 @@ class InstantMessage extends PureComponent {
       this.getRecipientsLabels(getRecipients(message));
   }
 
+  handleMessageDelete = messageDeleteHandler(this)
+  handleToggleMarkAsRead = toggleMarkAsReadHandler(this)
+  handleReply = replyHandler(this)
+
+  renderActions() {
+    const {
+      i18n,
+      message,
+      noInteractions, onOpenTags,
+    } = this.props;
+
+    if (noInteractions) {
+      return null;
+    }
+
+    return (
+      <div className="m-instant-message-aside__actions">
+        <Button
+          onClick={this.handleReply}
+          icon="reply"
+          title={i18n._('message-list.message.action.reply', null, { defaults: 'Reply' })}
+        />
+        <Button
+          onClick={onOpenTags}
+          icon="tags"
+          title={i18n._('message-list.message.action.tags', null, { defaults: 'Tags' })}
+        />
+        <Confirm
+          onConfirm={this.handleMessageDelete}
+          title={(<Trans id="message-list.message.confirm-delete.title">Delete a message</Trans>)}
+          content={(<Trans id="message-list.message.confirm-delete.content">The deletion is permanent, are you sure you want to delete this message ?</Trans>)}
+          render={confirm => (
+            <Button
+              onClick={confirm}
+              icon="trash"
+              title={i18n._('message-list.message.action.delete', null, { defaults: 'Delete' })}
+            />
+          )}
+        />
+        <Button
+          onClick={this.handleToggleMarkAsRead}
+          icon={message.is_unread ? 'envelope-open' : 'envelope'}
+          title={message.is_unread ?
+            i18n._('message-list.message.action.mark_as_read', null, { defaults: 'Mark as read' }) :
+            i18n._('message-list.message.action.mark_as_unread', null, { defaults: 'Mark as unread' })}
+        />
+      </div>
+    );
+  }
+
   render() {
     const {
       isLocked, encryptionStatus,
@@ -97,12 +161,16 @@ class InstantMessage extends PureComponent {
           <Icon type={this.getProtocolIconType(message)} />
           <Moment className="m-instant-message-author__time" format="HH:mm">{message.date}</Moment>
         </header>
-        <aside className="m-instant-message__info">
-          <div className="m-instant-message__participants m-instant-message-participants">
-            <TextBlock className="m-instant-message-participants__from">{author.label}</TextBlock>
-            <TextBlock className="m-instant-message-participants__to">{this.getRecipientsString(true)}<Icon type="caret-down" title={this.getRecipientsString(false)} /></TextBlock>
+        <aside className="m-instant-message__info m-instant-message-aside">
+          <div className="m-instant-message-aside__info">
+            <div className="m-instant-message__participants m-instant-message-participants">
+              <TextBlock className="m-instant-message-participants__from">{author.label}</TextBlock>
+              <TextBlock className="m-instant-message-participants__to">{this.getRecipientsString(true)}<Icon type="caret-down" title={this.getRecipientsString(false)} /></TextBlock>
+            </div>
+            <MessagePi illustrate={false} describe={false} message={message} />
           </div>
-          <MessagePi illustrate={false} describe={false} message={message} />
+          {this.renderActions()}
+          <TagList className="m-instant-message-aside__tags" message={message} />
         </aside>
         {
           isLocked ?