+++ src/frontend/web_application/src/scenes/Discussion/components/TagList/index.jsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { withI18n } from '@lingui/react';
+import { compose } from 'redux';
+import classnames from 'classnames';
+import { Badge } from '../../../../components';
+import { getTagLabelFromName, withTags } from '../../../../modules/tags';
+
+import './style.scss';
+
+const TagList = ({
+  className, i18n, tags: allTags, message,
+}) => {
+  if (!message.tags || message.tags.length === 0) {
+    return null;
+  }
+
+  return (
+    <ul className={classnames(className, 'm-message-tags-list')}>
+      {message.tags.map(tag => (
+        <li key={tag} className="s-mail-message__tag m-message-tags-list__tag">
+          <Badge>{getTagLabelFromName(i18n, allTags, tag)}</Badge>
+        </li>
+      ))}
+    </ul>
+  );
+};
+
+TagList.propTypes = {
+  i18n: PropTypes.shape({}).isRequired,
+  className: PropTypes.string,
+  message: PropTypes.shape({}).isRequired,
+  tags: PropTypes.arrayOf(PropTypes.shape({})),
+};
+TagList.defaultProps = {
+  className: undefined,
+  tags: [],
+};
+
+export default compose(
+  withTags(),
+  withI18n()
+)(TagList);