+++ src/frontend/web_application/src/scenes/Discussion/services/messageDeleteHandler.js
@@ -1,0 +1,5 @@
+export const messageDeleteHandler = ctx => () => {
+  const { message, onMessageDelete } = ctx.props;
+
+  onMessageDelete({ message });
+};