+++ src/frontend/web_application/src/scenes/Discussion/services/replyHandler.js
@@ -1,0 +1,5 @@
+export const replyHandler = ctx => () => {
+  const { onReply, message, push } = ctx.props;
+  onReply({ message });
+  push({ hash: 'reply' });
+};