+++ src/frontend/web_application/src/scenes/Discussion/services/toggleMarkAsReadHandler.js
@@ -1,0 +1,9 @@
+export const toggleMarkAsReadHandler = ctx => () => {
+  const { message, onMessageRead, onMessageUnread } = ctx.props;
+
+  if (message.is_unread) {
+    onMessageRead({ message });
+  } else {
+    onMessageUnread({ message });
+  }
+};